checks for upload file size
It seems that you can only check after the file has been submitted. This
might be bad if you upload 1gig and then after half an hour learn
that it is so large, but there seem to be no other (easier) way
unless you use a Java or flash uploader.